requirements.txt
SimpleCV
numpy
scipy
simpleOSC
